requirements.txt
sanic~=24.12.0
loguru~=0.7.3
aiohttp~=3.11.10
pycryptodome~=3.21.0
requests~=2.32.3
aiofiles~=24.1.0
python-dotenv~=1.0.1